refactor(web): extract hooks to comply with UI/Logic Separation
CLAUDE.md の MANDATORY ルール「UI/Logic Separation」に違反して
`useQuery` / `useMutation` / `useQueryClient` をコンポーネント内で
直接呼んでいた 9 ファイルを、それぞれ専用の hook に切り出した。

- dashboard widgets:
  - `use-active-session-widget` (cash/tournament queries + ticker)
  - `use-currency-balance-widget` + `useCurrencyBalanceOptions`
  - `use-recent-sessions-widget`
  - `use-summary-stats-widget`
- live-sessions/components/assign-ring-game-dialog:
  - `use-assign-ring-game` (stores + ringGames queries, assign/create mutations,
    invalidation via `invalidateTargets`, form state)
- stores/components:
  - `use-ring-game-form` (currencies query + useForm)
  - `use-tournament-form` (currencies query + useForm)
  - `use-tournament-tab` (tournaments + edit blind levels + create/update
    mutations + dialog state) / `useBlindStructureSummary`
- update-notes/components/update-notes-sheet:
  - `use-update-notes-viewed` (viewed list query + markViewed mutation +
    optimistic state)

コンポーネントは Props / hook 戻り値で受け取ったデータとハンドラを
描画するだけの Props 駆動の View に整理。
